+++ src/test/scala/za/co/absa/spark/commons/implicits/DataFrameImplicitsSuite.scala
@@ -22,6 +22,8 @@ import org.scalatest.funsuite.AnyFunSuite
 import za.co.absa.spark.commons.test.SparkTestBase
 
 class DataFrameImplicitsSuite extends AnyFunSuite with SparkTestBase  {
+  override implicit def sparkType: String = "local"
+
   import spark.implicits._
 
   private val columnName = "data"
